Fixed ---> Mini Game Initial Pos Bug

+++ Assets/Scripts/Runtime/Keys/SaveGameDataParams.cs
@@ -3,7 +3,7 @@
     public struct SaveGameDataParams
     {
         public int Level;
-        public float Money;
+        public int Money;
         public int IncomeLevel;
         public int StackLevel;
     }
